+++ packages/clerk-js/src/core/resources/internal.ts
@@ -29,6 +29,7 @@ export * from './OrganizationSuggestion';
 export * from './SamlAccount';
 export * from './Session';
 export * from './Passkey';
+export * from './ProtectConfig';
 export * from './PublicUserData';
 export * from './SessionWithActivities';
 export * from './SignIn';